requirements.txt
certifi==2017.7.27.1
colorama==0.3.9
future==0.16.0
libkeepass==0.2.0
lxml==4.1.0
peewee==2.10.2
pycrypto==2.6.1
python-telegram-bot==8.1.1
